+++ tools/rulelint/testdata/missing_subtest/pkg/somefile_test.go
@@ -1,0 +1,9 @@
+package pkg
+
+import "testing"
+
+func TestSafety(t *testing.T) {
+	t.Run("clamp_below_min", func(t *testing.T) {
+		// only this subtest exists
+	})
+}